purgador/requirements.txt
hikari
hikari-lightbulb
uvloop; sys_platform != 'win32'
python-dotenv
lavaplayer